Added OSM parsing

+++ src/PathFinding.cpp
@@ -3,19 +3,19 @@
 #include <cstdint>
 #include <vector>
 
-bool Djikstra(Graph& graph, uint32_t start_node, uint32_t end_node, std::vector<uint32_t>& out_path)
+bool Djikstra(Graph& graph, uint64_t start_node, uint64_t end_node, std::vector<uint64_t>& out_path)
 {
-    std::vector<double>    dist(graph.nodes.size(), INFINITY);
-    std::vector<uint32_t> prev(graph.nodes.size(), 0xFFFFFFFF);
-    std::vector<uint32_t> queue = {start_node};
+    std::vector<double>   dist(graph.nodes.size(), INFINITY);
+    std::vector<uint64_t> prev(graph.nodes.size(), 0xFFFFFFFF);
+    std::vector<uint64_t> queue = {start_node};
 
     dist[start_node] = 0;
 
-    std::vector< std::vector<uint32_t> > adj_list(graph.nodes.size());
+    std::vector< std::vector<uint64_t> > adj_list(graph.nodes.size());
     for (const Edge& edge : graph.edges) 
     {
         // Undirected graph, so add both directions (if we want directed, we'd only add one)
-        for (uint32_t i = 0; i < edge.nodeRefs.size() - 1; i++)
+        for (uint64_t i = 0; i < edge.nodeRefs.size() - 1; i++)
         {
             adj_list[edge.nodeRefs[i]].push_back(edge.nodeRefs[i + 1]);
             adj_list[edge.nodeRefs[i + 1]].push_back(edge.nodeRefs[i]);
@@ -25,8 +25,8 @@ bool Djikstra(Graph& graph, uint32_t start_node, uint32_t end_node, std::vector<
     while (!queue.empty()) 
     {
         // Find the node in the queue with the smallest distance
-        uint32_t current = queue[0];
-        for (uint32_t node : queue) 
+        uint64_t current = queue[0];
+        for (uint64_t node : queue) 
         {
             if (dist[node] < dist[current]) {
                 current = node;
@@ -50,7 +50,7 @@ bool Djikstra(Graph& graph, uint32_t start_node, uint32_t end_node, std::vector<
         queue.erase(std::remove(queue.begin(), queue.end(), current), queue.end());
 
         // Update distances to neighbors
-        for (uint32_t neighbor : adj_list[current]) 
+        for (uint64_t neighbor : adj_list[current]) 
         {
             double alt = dist[current] + sqrt(pow(graph.nodes[current].x - graph.nodes[neighbor].x, 2) + pow(graph.nodes[current].y - graph.nodes[neighbor].y, 2));
             if (alt < dist[neighbor])
