+++ src/PathFinding.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "Graph.h"
 
-bool Djikstra(Graph& graph, uint32_t start_node, uint32_t end_node, std::vector<uint32_t>& out_path);
+bool Djikstra(Graph& graph, uint64_t start_node, uint64_t end_node, std::vector<uint64_t>& out_path);
